Hexagonal Architecture Example in Java and Spring
