--- spring-hexagonal/README.md
@@ -1,6 +1,0 @@
-=========
-### A Quick and Practical Example of Hexagonal Architecture in Java
-
-### To test
-- Start application as spring boot 
-- Test api - http://localhost:8080/users
