--- spring-hexagonal/application/src/main/java/com/baeldung/hexagonal/SpringHexagonalApplication.java
@@ -1,12 +1,0 @@
-package com.baeldung.hexagonal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringHexagonalApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringHexagonalApplication.class, args);
-	}
-}
