--- spring-hexagonal/domain/src/main/java/com/baeldung/hexagonal/domain/ports/GetUserService.java
@@ -1,8 +1,0 @@
-package com.baeldung.hexagonal.domain.ports;
-
-import com.baeldung.hexagonal.domain.User;
-import java.util.List;
-
-public interface GetUserService {
-    List<User> getAllUsers();
-}
