+++ spring-hexagonal/src/main/java/com/baeldung/app/SpringHexagonalApplication.java
@@ -1,0 +1,13 @@
+package com.baeldung.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringHexagonalApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringHexagonalApplication.class, args);
+    }
+
+}
