+++ spring-hexagonal/src/main/java/com/baeldung/app/domain/User.java
@@ -1,0 +1,30 @@
+package com.baeldung.app.domain;
+
+public class User {
+
+    private int id;
+    private String name;
+
+    public User(int id, String name) {
+        super();
+        this.id = id;
+        this.name = name;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
