+++ spring-hexagonal/src/main/java/com/baeldung/app/domain/ports/UserService.java
@@ -1,0 +1,9 @@
+package com.baeldung.app.domain.ports;
+
+import java.util.List;
+
+import com.baeldung.app.domain.User;
+
+public interface UserService {
+    List<User> getUsers();
+}
