+++ spring-hexagonal/src/main/java/com/baeldung/app/rest/adapters/UserController.java
@@ -1,0 +1,27 @@
+package com.baeldung.app.rest.adapters;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.baeldung.app.domain.User;
+import com.baeldung.app.domain.ports.UserService;
+
+@RestController
+public class UserController {
+
+    private UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @RequestMapping(value = "/users", method = RequestMethod.GET)
+    public List<User> getUsers() {
+        return userService.getUsers();
+    }
+}
